requirements.txt
# Core dependencies
aiohttp>=3.8.0
asyncio-timeout>=4.0.0;python_version<"3.11"
PyYAML>=6.0

# Testing dependencies
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-mock>=3.10.0

# Development dependencies
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0